+++ src/app.js
@@ -2,49 +2,58 @@
 
 	var curr;
     mbrApp.regExtension({
-        name: "anchor-editor",
+        name: "witsec-anchor-editor",
         events: {		
-            load: function() {
-                var a = this;
-				a.$body.find(".navbar-devices").append('<li class="witsec-anchor-editor-show" style="width:66px; height:58px; cursor:pointer" data-tooltipster="bottom" title="Anchor Editor"><i class="mbr-icon-link mbr-iconfont"></i></li>');
-				a.$body.on("click", ".witsec-anchor-editor-show", function(b) {
-					var currentPage = mbrApp.Core.currentPage;		
-					var complist = '';
-					var curr = mbrApp.Core.resultJSON[currentPage].components[0];
-					for (x in mbrApp.Core.resultJSON[currentPage].components){
-						compname = mbrApp.Core.resultJSON[currentPage].components[x]._name
-						complist = complist + '<option value="'+ x +': '+compname +'">'+ x +': '+ compname + '</option>';
-					}
+            beforeAppLoad: function() {
+                mbrApp.Core.addFilter("prepareComponent", function(a, b) {
+					// 'a' is the component window's HTML as string. We need to jQuery that, so we can do magic stuff with it
+					var h = jQuery(a);
 
-					if (complist == "") {
-						mbrApp.alertDlg("You have to add at least one block before you can use the Anchor Editor.");
+					// Add edit button to component buttons
+					var btn = '<span class="mbr-btn mbr-btn-default mbr-icon-link witsec-anchor-editor-shortcut" data-tooltipster="bottom" title="Edit Anchor"></span><style>.witsec-anchor-editor-shortcut:hover { background-color: #42a5f5 !important; }</style>';
+					if (h.find(".component-params").length)
+						h.find(".component-params").before(btn);
+					else if (h.find(".component-remove").length)
+						h.find(".component-remove").before(btn);
+
+					// Get the HTML as a string, then return that
+					a = h.prop("outerHTML");
+					h.remove();
+
+					return a;
+				});
+			},
+
+			load: function() {
+                var a = this;
+
+				a.$template.on("click", ".witsec-anchor-editor-shortcut", function(e) {
+					// Find the index of the clicked icon
+					a.$template.find('.witsec-anchor-editor-shortcut').each(function(index, obj) {
+						if (e.target == obj) {
+							curr = mbrApp.Core.resultJSON[mbrApp.Core.currentPage].components[index];
+						}
+					});
+
+					// If curr is null, something is wrong
+					if (curr === null) {
+						mbrApp.alertDlg("An error occured while opening the Anchor Editor.");
 						return false;
 					}
 
-					mbrApp.$body.on("change", "#witsec-anchor-editor-sections", function(a) {
-						var ind = this.selectedIndex;
-						var currentPage = mbrApp.Core.currentPage;
-						curr = mbrApp.Core.resultJSON[currentPage].components[ind];
-						$("#witsec-anchor-editor-name").val(curr._anchor);
-					});
-
 					mbrApp.showDialog({
-						title: "Anchor Editor",
+						title: "Edit Anchor Name",
 						className: "",
 						body: [
-							'<div>Make sure you only use valid characters in your anchor name.</div>',
-							'<div>&nbsp;</div>',
-							'<div class="row">',
-							'  <div class="col-lg-5"><select class"form-control" id="witsec-anchor-editor-sections">', complist, '</select></div>',
-							'  <div class="col-lg-7">Anchor: <input type="text" class"form-control" id="witsec-anchor-editor-name" value="', curr._anchor, '"></div>',
-							'</div>'
+							'<div>Anchor: <input type="text" class"form-control" id="witsec-anchor-editor-name" value="', curr._anchor, '"></div>'
 						].join("\n"),
 						buttons: [{
 							label: "SAVE",
 							default: !0,
 							callback: function() {
 								try {
-									var anchor = $('#witsec-anchor-editor-name').val();
+									// Replace any whitespace characters with a dash, any other character is allowed in HTML5
+									var anchor = $('#witsec-anchor-editor-name').val().replace(/\s/g, "-");
 
 									// Check if the entered value isn't empty
 									if (anchor.trim() == "") {
@@ -60,20 +69,13 @@
 									mbrApp.runSaveProject();
 								}
 								catch(err){
-									mbrApp.alertDlg(err.name + ' with message : ' +err.message);
+									mbrApp.alertDlg(err.name + ': ' + err.message);
 								}
 							}
-						},
-							{
-							label: "CLOSE",
-							default: !0,
-							callback: function() {									
-							}
-						}
-						]
+						}]
 					});
 				});				
-            }
+			}
         }
     })
 })(jQuery, mbrApp);