+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## 0.2 (2019-03-29)
+
+- Removed the icon in the navigation bar
+- Added anchor icon to the component's window, so it's easier to see what section's anchor you're editing
+- Added simple regex to replace whitespace characters with a dash
+
 ## 0.1 (2019-03-25)
 
 This is the initial release of the Anchor Editor extension for Mobirise.